pyproject.toml
[project]
name = "inventory-and-sales-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.13.2",
    "flasgger>=0.9.7.1",
    "flask==3.0.3",
    "gunicorn==21.2.0",
    "marshmallow==3.21.1",
    "psycopg[binary]==3.2.1",
    "pytest==8.3.3",
    "pytest-dotenv==0.5.2",
    "pytest-flask==1.3.0",
    "pytest-mock==3.14.0",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.34",
]
